requirements.txt
pandas
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9